Headless mode functionnal
The headless mode of the visio finder is functionnal.

Currently, it can read a config file to search for specific targets (file or folder) and create shortcuts for those targets. It supports regular expressions for the target name and for filtering the folders to search.

To launch in headless mode: `Visiofinder.exe --headless --config path/to/config/file`